pyproject.toml
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "exa-haystack"
version = "0.1.0"
description = "Haystack integration for Exa search API"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [{ name = "Exa", email = "support@exa.ai" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "haystack-ai>=2.0.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "types-requests",
]

[project.urls]
Homepage = "https://github.com/exa-labs/exa-haystack"
Documentation = "https://docs.exa.ai"
Repository = "https://github.com/exa-labs/exa-haystack"

[tool.hatch.build.targets.wheel]
packages = ["src/haystack_integrations"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
exclude = ["lazy_imports"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
